requirements.txt
flask>=3.0.0
pymongo>=4.6.0
requests>=2.31.0
psutil>=5.9.0
pyautogui>=0.9.53
opencv-python>=4.8.0
mss>=8.0.0
numpy>=1.26.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
pynput>=1.7.6
pillow>=10.0.0
python-dotenv>=1.0.0
